move LayoutMenu components out of Geral subfolder
